Use size_t for prefRowsPerColumn

+++ src/openrct2-ui/interface/Dropdown.h
@@ -36,10 +36,10 @@ namespace OpenRCT2::Ui::Windows
 
     void WindowDropdownShowText(
         const ScreenCoordsXY& screenPos, int32_t extray, ColourWithFlags colour, uint8_t flags, size_t num_items,
-        int8_t prefRowsPerColumn = 0);
+        size_t prefRowsPerColumn = 0);
     void WindowDropdownShowTextCustomWidth(
         const ScreenCoordsXY& screenPos, int32_t extray, ColourWithFlags colour, uint8_t custom_height, uint8_t flags,
-        size_t num_items, int32_t width, int8_t prefRowsPerColumn = 0);
+        size_t num_items, int32_t width, size_t prefRowsPerColumn = 0);
 
     void WindowDropdownShowImage(
         int32_t x, int32_t y, int32_t extray, ColourWithFlags colour, uint8_t flags, int32_t numItems, int32_t itemWidth,
